move types n OracleConfigurationExtension.cs to their own file #39
